+++ common/TracyQueue.hpp
@@ -32,14 +32,14 @@ struct QueueZoneBegin
     int64_t time;
     uint64_t thread;
     uint64_t srcloc;    // ptr
-    int8_t cpu;
+    uint32_t cpu;
 };
 
 struct QueueZoneEnd
 {
     int64_t time;
     uint64_t thread;
-    int8_t cpu;
+    uint32_t cpu;
 };
 
 struct QueueStringTransfer
